requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
discord.py==2.4.0
env==0.1.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
multidict==6.0.5
playwright==1.46.0
pyee==11.1.0
pyotp==2.9.0
python-dotenv==1.0.1
typing_extensions==4.11.0
yarl==1.9.4
